requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-huggingface
openai
tiktoken
faiss-cpu
pydantic
python-dotenv
sentence-transformers